Add solution for 209 Minimum Size Subarray Sum in Java
- The solution employs the sliding window technique to find the smallest subarray whose sum is greater than or equal to a target value.
- O(N) Time Complexity. Even though we are using a while loop inside a for loop, it is still linear time because the
inner loop is not executing n times for every iteration of the outer loop.
Source - LeetCode: "A sliding window guarantees a maximum of 2n window iterations. This is what is referred to as amortized analysis - even though the worst
case for an iteration inside the for loop is O(n)."
- O(1) Space Complexity.